Refactor CMake var _dgl_library -> _dgl_has_ui in plugin build functions
Plugin build functions (dpf__build_<PLUGIN_TYPE>) invokes
dpf__add_ui_main() which has a parameter HAS_UI. This parameter acts
like a switch, controlling if DistrhoPluginMain.cpp shall be compiled.

Before this patch, value of _dgl_library is passed into HAS_UI. However,
this will make it ambiguous. Variable _dgl_library should only be served
as a flag of DGL UI type, and should not be a switch of whether to build
DistrhoPluginMain.cpp or not.

What's more, since DPF's CMake build system starts to support external
UI, which is not limited to DGL, simply checking _dgl_library for
dgl__add_ui_main() is no longer relevant.

So, instead, I use variable _dgl_has_ui which keeps to the point. It
will be set to ON if _dgl_library is non-empty or _dgl_external is ON.