Fix crash on CLAP hosts that do not implement latency extension
When DISTRHO_PLUGIN_WANT_LATENCY is defined, CLAP plugins assume the host implements the latency CLAP extension and do not check whether the pointer to the host's latency extension is nullptr. As a result, a null pointer dereference can occur when the host proceeds to use the plugin. This commit makes plugins that require the latency extension to check that hosts actually implement it, and if this requirement is not met, plugin initialization will fail.